Wrap spells page search params in suspense
